Gemini_API/requirements.txt
flask
google-genai
Pillow
python-dotenv==1.0.0